requirements.txt
webcolors==1.3
urllib3==1.22
requests==2.18.4
Pillow==6.2.0
beautifulsoup4==4.8.1
